requirements.txt
joblib==1.5.3
numpy==2.4.4
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.1
threadpoolctl==3.6.0
